+++ .gitignore
@@ -1,3 +1,4 @@
+.tmp/
 node_modules/
 data/
 __pycache__
