+++ envs/blue
@@ -1,4 +1,4 @@
 API_PORT=18080
-APP_ENV=prod
+APP_ENV=blue
 ENABLE_CACHE=true
 ENVIRONMENT=prod
