+++ envs/green
@@ -1,4 +1,4 @@
-API_PORT=18080
-APP_ENV=prod
+API_PORT=18081
+APP_ENV=green
 ENABLE_CACHE=true
 ENVIRONMENT=prod
